Polyfill the incremental generator ForAttributeWithMetadataName from roslyn (for JsonGenerator).
